requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.55.3
httpx==0.27.2
anyio==3.7.1
pydantic==2.7.4
pydantic-core==2.18.4
python-dotenv==1.0.0
Pillow>=10.3.0
streamlit==1.39.0
requests>=2.31.0
